Add Async Vdp1 renderer - It is sync for the moment...

+++ yabause/src/patternManager.c
@@ -8,8 +8,11 @@ Pattern* patternCache[0xFFFF];
 
 #define CACHE_LIFETIME 10;
 
-static void deleteCachePattern(Pattern* pat) {
+void deleteCachePattern(Pattern* pat) {
 	if (pat == NULL) return;
+	pat->inUse--;
+	if (pat->inUse != 0) return;
+	pat->managed = 0;
 	glDeleteTextures(1, &pat->tex);
 	free(pat);
 }
@@ -41,6 +44,8 @@ void recycleCache() {
 	for (i = 0; i < 0xFFFF; i++) {
 		Pattern* tmp = patternCache[i];
 		if (tmp == NULL) continue;
+		tmp->inUse = 0;
+		tmp->managed = 1;
 		tmp->frameout--;
 		if (tmp->frameout == 0) {
 			deleteCachePattern(tmp);
@@ -53,18 +58,21 @@ Pattern* getCachePattern(int param0, int param1, int param2, int w, int h) {
   Pattern *pat = patternCache[getHash(param0, param1)];
   if ((pat!= NULL) && (pat->param[0]==param0) && (pat->param[1]==param1) && (pat->param[2]==param2) && (pat->width == w) && (pat->height == h)) {
         pat->frameout = CACHE_LIFETIME;
+	pat->inUse++;
   	return pat;
   } else {
 	return NULL;
   }
 }
-int size = 0;
 
 void addCachePattern(Pattern* pat) {
 	Pattern *collider = patternCache[getHash(pat->param[0], pat->param[1])];
+	if ((collider != NULL) && (collider->inUse > 0)) return;
 	if (collider != NULL) {
 		deleteCachePattern(collider);
 	}
+	pat->managed = 1;
+	pat->inUse = 1;
 	patternCache[getHash(pat->param[0], pat->param[1])] = pat;
 }
 
@@ -75,6 +83,8 @@ Pattern* createCachePattern(int param0, int param1, int param2, int w, int h, fl
 	new->param[2] = param2;
         new->width = w;
 	new->height = h;
+	new->managed = 0;
+	new->inUse = 0;
         new->tw = tw;
         new->th = th;
 	new->mesh = mesh;
